requirements.txt
streamlit
faiss-cpu
google-generativeai
sentence-transformers
PyPDF2
numpy
